requirements-worker.txt
aiokafka>=0.8.0
aio-pika>=9.3.0
httpx>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-instrumentation-sqlalchemy